LayoutTests/imported/w3c/web-platform-tests/css/css-writing-modes/bidi-line-break-001-expected.txt


PASS
    <span class="isolate" dir="ltr">00</span> <span class="isolate" dir="ltr">00</span>

PASS
    <span class="embed" dir="ltr">00</span> <span class="embed" dir="ltr">00</span>

PASS
    <span dir="ltr">00</span> <span dir="ltr">00</span>

PASS
    <bdi dir="ltr">00</bdi> <bdi dir="ltr">00</bdi>

PASS
    <bdo dir="ltr">00</bdo> <bdo dir="ltr">00</bdo>

PASS
    <span class="isolate" dir="ltr">00 </span><span class="isolate" dir="ltr">00</span>

PASS
    <span class="embed" dir="ltr">00 </span><span class="embed" dir="ltr">00</span>

00 00
00 00
00 00
00 00
00 00
00 00
00 00
